Update to stratum.js
I got this handed to me by zone117x to enable the haproxy to send the real user ip using the send-proxy flag. He will re-work it and integrate it better, but I've pushed this commit for now for those who want to have it working.

Also, I fixed an issue where miners using Uppercase characters in their workers wouldn't get authed by the stratum. I simply added .toLowerCase(); where the stratum looks in the db for the workername.